Dev temp date range

+++ temperature/views.py
@@ -4,6 +4,7 @@ from rest_framework.permissions import IsAdminUser
 from rest_framework.response import Response
 from .models import Temperature
 from .serializers import TemperatureSerializer
+from datetime import datetime
 
 
 @api_view(['GET', 'POST'])
@@ -12,15 +13,20 @@ def temperature_list(request):
         temperatures = Temperature.objects.all()
         serializer = TemperatureSerializer(temperatures, many=True)
         return Response(serializer.data)
-    
-    elif request.method == 'POST':
-        serializer = TemperatureSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.erros, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['GET'])
+    elif request.method == 'POST':
+        if request.user.is_superuser:
+            serializer = TemperatureSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        elif not request.user.is_superuser:
+            return Response("Unauthorized Request", status=status.HTTP_401_UNAUTHORIZED)
+    return Response(serializer.erros, status=status.HTTP_400_BAD_REQUEST)
+   
+
+
+@api_view(['GET', 'DELETE'])
 def temperature_detail(request, pk):
     try:
         temperature = Temperature.objects.get(pk=pk)
@@ -29,24 +35,18 @@ def temperature_detail(request, pk):
     if request.method == 'GET':
         serializer = TemperatureSerializer(temperature)
         return Response(serializer.data)
+    elif request.method == 'DELETE' and request.user.is_superuser:
+            temperature.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+    return Response("Unauthorized Request", status=status.HTTP_401_UNAUTHORIZED)
 
 
-@api_view(['DELETE'])
-@permission_classes([IsAdminUser])
-def temperature_detail(request, pk):
-    try:
-        temperature = Temperature.objects.get(pk=pk)
-    except Temperature.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-    if request.method == 'DELETE':
-       temperature.delete()
-       return Response(status=status.HTTP_204_NO_CONTENT)
-# from rest_framework import generics
-# from rest_framework import serializers
-# from rest_framework.permissions import IsAdminUser
-# from .models import Temperature
-# from .serializers import TemperatureSerializer
-
-# # class Temperature_list(generics.ListCreateAPIView):
-# #     queryset = Temperature.objects.all()
-# #     serializer_class = TemperatureSerializer
+@api_view(['GET'])
+def temperature_list_date_range(request):
+    if request.method == 'GET':
+        start = request.GET.get('start')
+        end = request.GET.get('end')
+        temperatures = Temperature.objects.filter(
+            date__range=[start, end])
+        serializer = TemperatureSerializer(temperatures, many=True)
+        return Response(serializer.data)
